refactor: Create dir for ChildPanelExportReport files
